🚀 Feature: Split app into smaller components and hooks
