+++ src/hooks/useCurrencyRates.ts
@@ -1,0 +1,55 @@
+import { useContext, useEffect, useState } from "react";
+import { ConverterContext } from "../feature/converter/ConverterContext.tsx";
+
+const useCurrencyRates = () => {
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState<string>("");
+
+  const context = useContext(ConverterContext);
+  if (!context) {
+    throw new Error("Converter must be used within a ConverterProvider");
+  }
+
+  const { currencyRates, updatedAt, setCurrencyRates, setUpdatedAt } = context;
+  const dataExists = !!currencyRates;
+
+  useEffect(() => {
+    fetchData();
+  }, []);
+
+  const fetchData = async () => {
+    try {
+      setIsLoading(true);
+
+      const res = await fetch(
+        `https://v6.exchangerate-api.com/v6/a2b1eb34c3500f20de9a5727/latest/USD`,
+      );
+      const data = await res.json();
+
+      const rates = {
+        USD: 1,
+        UZS: data.conversion_rates.UZS as number,
+        EUR: data.conversion_rates.EUR as number,
+      };
+
+      setCurrencyRates(rates);
+      setUpdatedAt(data.time_last_update_utc);
+      setError("");
+    } catch (e) {
+      const message = e instanceof Error ? e.message : "Unknown error";
+      setError(`Failed to update rates: ${message}`);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return {
+    dataExists,
+    isLoading,
+    error,
+    fetchData,
+    updatedAt,
+  };
+};
+
+export default useCurrencyRates;
